+++ packages/navbar/src/navbar-item.tsx
@@ -1,15 +1,25 @@
 import React from 'react';
 
-import { NavbarItemWrapper } from './private';
+import { ThemeContext } from '@uireact/foundation';
 
-export interface UiNavbarItemProps {
-  children?: React.ReactNode;
-  /** Alignment where the item content will render - CENTER Default */
-  align?: 'right' | 'center' | 'left';
-}
+import styled from 'styled-components';
 
-export const UiNavbarItem: React.FC<UiNavbarItemProps> = ({ children, align }: UiNavbarItemProps) => {
-  return <NavbarItemWrapper align={align}>{children}</NavbarItemWrapper>;
+import { UiNavbarItemProps, privateNavbarItemProps } from './types';
+
+const Div = styled.div<privateNavbarItemProps>`
+  ${(props) => `
+    ${!props.active ? 'background: transparent !important;' : ''}
+  `}
+`;
+
+export const UiNavbarItem: React.FC<UiNavbarItemProps> = ({ active, children }: UiNavbarItemProps) => {
+  const themeContext = React.useContext(ThemeContext);
+
+  return (
+    <Div active={active} customTheme={themeContext.theme} selectedTheme={themeContext.selectedTheme}>
+      {children}
+    </Div>
+  );
 };
 
 UiNavbarItem.displayName = 'UiNavbarItem';