+++ packages/navbar/src/types/navbar-item-props.ts
@@ -1,0 +1,9 @@
+import { UiReactElementProps, UiReactPrivateElementProps } from '@uireact/foundation';
+
+export type UiNavbarItemProps = {
+  children?: React.ReactNode;
+  /** If the current option is active */
+  active?: boolean;
+} & UiReactElementProps;
+
+export type privateNavbarItemProps = UiNavbarItemProps & UiReactPrivateElementProps;