+++ packages/text/src/ui-link.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import { Link } from 'react-router-dom';
+
 import styled from 'styled-components';
 
 import { TextSize, ThemeContext, getColorCategory, getTextSize, getThemeStyling } from '@uireact/foundation';
@@ -11,6 +13,7 @@ const Anchor = styled.a<privateLinkProps>`
   ${(props) => `
     ${getThemeStyling(props.customTheme, props.selectedTheme, getDynamicLinkMapper(getColorCategory(props.category)))}
     font-size: ${getTextSize(props.customTheme, props.size || TextSize.regular)};
+    ${props.fullWidth ? 'width: 100%; display: inline-block;' : ''}
   `}
 
   cursor: pointer;
@@ -22,23 +25,68 @@ const Anchor = styled.a<privateLinkProps>`
   }
 `;
 
+const StyledLinkWrapper = styled.span<privateLinkProps>`
+  a {
+    ${(props) => `
+      ${getThemeStyling(props.customTheme, props.selectedTheme, getDynamicLinkMapper(getColorCategory(props.category)))}
+      font-size: ${getTextSize(props.customTheme, props.size || TextSize.regular)};
+    `}
+
+    cursor: pointer;
+    outline: none;
+    text-decoration: none;
+
+    :hover {
+      text-decoration: underline;
+    }
+  }
+`;
+
 export const UiLink: React.FC<UiLinkProps> = ({
   theme = 'secondary',
   children,
   handleClick,
   href,
+  fullWidth,
   ref,
   referrerpolicy,
   size,
   target,
+  useReactLink,
+  testId,
 }: UiLinkProps) => {
   const themeContext = React.useContext(ThemeContext);
+
+  if (useReactLink && href) {
+    return (
+      <StyledLinkWrapper
+        category={theme}
+        customTheme={themeContext.theme}
+        fullWidth={fullWidth}
+        onClick={handleClick}
+        selectedTheme={themeContext.selectedTheme}
+        size={size}
+      >
+        <Link
+          to={href}
+          data-testid={testId}
+          role="link"
+          target={target}
+          ref={ref}
+          referrerPolicy={referrerpolicy || 'no-referrer'}
+        >
+          {children}
+        </Link>
+      </StyledLinkWrapper>
+    );
+  }
 
   return (
     <Anchor
       category={theme}
       customTheme={themeContext.theme}
       href={href}
+      fullWidth={fullWidth}
       onClick={handleClick}
       ref={ref}
       referrerpolicy={referrerpolicy}
@@ -46,6 +94,7 @@ export const UiLink: React.FC<UiLinkProps> = ({
       size={size}
       target={target}
       role="link"
+      data-testid={testId}
     >
       {children}
     </Anchor>