+++ packages/form/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.8](https://github.com/inavac182/ui-react/compare/@uireact/form@1.0.7...@uireact/form@1.0.8) (2023-06-08)
+
+**Note:** Version bump only for package @uireact/form
+
+
+
+
+
+## 1.0.7 (2023-06-07)
+
+**Note:** Version bump only for package @uireact/form
+
+
+
+
+
 ## [1.0.6](https://github.com/inavac182/ui-react/compare/@uireact/form@1.0.5...@uireact/form@1.0.6) (2023-05-31)
 
 **Note:** Version bump only for package @uireact/form